+++ src/TypeGuards/helpers/asTypeGuard.ts
@@ -1,6 +1,6 @@
 import type { Param0 } from '../../types/Func.ts'
 import type { Predicate } from '../../types/Predicate.ts'
-import type { CustomStruct } from '../../validators/schema/types/index.ts'
+import type { V3 } from '../../validators/schema/types/v3/index.ts'
 import type { TypeGuard } from '../types/index.ts'
 
 import { isRuleStruct } from '../../validators/schema/helpers/isRuleStruct.ts'
@@ -13,19 +13,19 @@ type OptionalizedKeys = 'rules'
 
 export function asTypeGuard<T>(
     predicate: Predicate<any>,
-    metadata?: Omit<CustomStruct<T>, OmittedKeys | OptionalizedKeys> &
-        Partial<Pick<CustomStruct<T>, OptionalizedKeys>>
+    metadata?: Omit<V3.CustomStruct<T>, OmittedKeys | OptionalizedKeys> &
+        Partial<Pick<V3.CustomStruct<T>, OptionalizedKeys>>
 ): TypeGuard<T>
 export function asTypeGuard<P extends Predicate<any>>(
     predicate: P,
-    metadata?: Omit<CustomStruct<Param0<P>>, OmittedKeys | OptionalizedKeys> &
-        Partial<Pick<CustomStruct<Param0<P>>, OptionalizedKeys>>
+    metadata?: Omit<V3.CustomStruct<Param0<P>>, OmittedKeys | OptionalizedKeys> &
+        Partial<Pick<V3.CustomStruct<Param0<P>>, OptionalizedKeys>>
 ): TypeGuard<Param0<P>>
 
 export function asTypeGuard<T>(
     predicate: Predicate<unknown>,
-    metadata: Omit<CustomStruct<T>, OmittedKeys | OptionalizedKeys> &
-        Partial<Pick<CustomStruct<T>, OptionalizedKeys>> = {}
+    metadata: Omit<V3.CustomStruct<T>, OmittedKeys | OptionalizedKeys> &
+        Partial<Pick<V3.CustomStruct<T>, OptionalizedKeys>> = {}
 ): TypeGuard<T> {
     if (!isUnaryFunction<TypeGuard>(predicate))
         throw new Error('predicate must be a unary function')
@@ -46,7 +46,7 @@ export function asTypeGuard<T>(
 
         type: 'custom',
         schema: predicate,
-    } as CustomStruct<T>
+    } as V3.CustomStruct<T>
 
     return setCustomStructMetadata<any>(struct, setAsTypeGuard(predicate))
 }
