+++ src/TypeGuards/helpers/isInstanceOf.ts
@@ -1,5 +1,5 @@
 import { setStructMetadata } from '../../validators/schema/helpers/setStructMetadata.ts'
-import type { V3 } from '../../validators/schema/types/index.ts'
+import type { V3 } from '../../validators/schema/types/v3/index.ts'
 import type { ConstructorSignature } from '../types/index.ts'
 import { __curry_param__ } from './constants.ts'
