+++ src/validators/schema/and.ts
@@ -1,7 +1,7 @@
 import type { TypeGuard } from '../../TypeGuards/types/index.ts'
 import type { Merge } from '../../types/index.ts'
 import type { Custom } from '../rules/types/index.ts'
-import type { V3 } from './types/index.ts'
+import type { V3 } from './types/v3/index.ts'
 import type {
     IntersectionSchema,
     IntersectionSchemaEntry,
