+++ src/validators/schema/helpers/setStructMetadata.ts
@@ -1,7 +1,7 @@
 import type Generics from '../../../Generics/index.ts'
 import type { TypeGuard } from '../../../TypeGuards/types/index.ts'
 import type { TypeFromArray } from '../../../types/index.ts'
-import type { V3 } from '../types/index.ts'
+import type { V3 } from '../types/v3/index.ts'
 
 import { setMetadata } from '../../../TypeGuards/helpers/setMetadata.ts'
 import { attachStandardSchema } from '../../standard-schema/attachStandardSchema.ts'
