+++ src/validators/schema/object.ts
@@ -1,7 +1,7 @@
 import type { TypeGuard } from '../../TypeGuards/types/index.ts'
 import type { Custom } from '../rules/types/index.ts'
 import type { Sanitize, NormalizedValidatorMap, ValidatorMap } from '../types/index.ts'
-import type { V3 } from './types/index.ts'
+import type { V3 } from './types/v3/index.ts'
 import type { ObjectSchema } from './types/ObjectSchema.ts'
 
 import { join } from '../../helpers/Experimental/join.ts'
