+++ src/validators/schema/types/FluentSchema.ts
@@ -2,7 +2,7 @@
 import type { TypeGuard } from '../../../TypeGuards/index.ts'
 import type { Fn, ThrowFn } from '../../../types/Func.ts'
 import type { Custom } from '../../rules/types/index.ts'
-import type { ValidateReturn } from '../../SchemaValidator.ts'
+import type { ValidateReturn } from '../../types/ValidateReturn.ts'
 import type { ValidationErrors } from '../../ValidationErrors.ts'
 import type { StandardSchemaV1 } from '../../standard-schema/types.ts'
 import type { FluentOptionalSchema } from './FluentOptionalSchema.ts'
